Altered Code to work with Adafruit M4 Metro
Yikes, that was an adventure. I arrived to try and show Squinky my code
on the new M4 boards and it was a lot more complicated to get it to work
on the M4 than expected. I had to edit Boards.h to include a definition
of the M4 -- I should probably include what I put in boards.h for future
reference. Also, I had to include junk code at the start of the file and
I had to disable the servo library. Yikes!