Enric helped with text display
So, in a nutshell, broke some arduino code to start getting the LCD
things working (they're not working yet),
but text display is really fancy and will scroll vertically thanks to
Enric.

Enric also helped me to figure out a better workflow for the information
handling, particularly with the tags, story, and audio files.

Will have to fix the arduino code and get the serial port LCD screen
side of things working! to be continued!