Update log054.ogg
changed the audio log to something less annoying for tests

also, the audio wasn't playing because I had my headphones in.

I am concerned about how this will play with the Raspberry Pi when I go headless though: soundjs.min.js:18 The AudioContext was not allowed to start. It must be resumed (or created) after a user gesture on the page. https://goo.gl/7K7WLu