More socket issues
Ben Gattet helped answer some of my socket questions but I don't know why the socket stuff is currently not working. I'll try and see if I can't fix it, but so far clearly we're not even making it into the chunk-related socket info because I'm not getting the debug messages on the console.