Update README.md
Sound output is working but it's possible that the Autoplay block on chromium may return to haunt me (hard to say).

BUT! the game is working on the Pi, so that's a big step.

Currently working on trying to get PM2 up and running so everything will just load when I restart the Pi.