Memory Leak KILLED
So that code actually worked.

Last problem to solve is getting the server to autostart on boot, but if I don't solve that before Tuesday, it is okay.

That means that it's time to craft the controllers and the sculptures! (after I install all of this on the last pi.)